Create Startup file and move firebase init to it
